[UPDATE] Adicionado suporte ao docker e implementado migração no código

+++ DataAccess/Product.cs
@@ -6,7 +6,7 @@ namespace ecommerce_product.DataAccess
         // DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int ProductID { get; set; }
         public string ProductName { get; set; }
-        public float ProductPrice { get; set; }
+        public double ProductPrice { get; set; }
         public string ProviderCNPJ { get; set; }
         public string Description { get; set; }
